+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### [Unreleased][unreleased]
 
+#### Added
+- Status command - list of migrations already executed and list of migrations to execute
+
 ### [0.5.0] - 2016-08-03
 #### Added
 - column types date, bigint, float